Added login functionality for admin user, most endpoints require you to be logged in as admin now. Added base.html for other templates to inherit from instead of retyping the same html in each template. Updated the move_to_main endpoint so that it doesn't change the station entry in the database unless the file is moved successfully and now moves based on file_path instead of title. Updated the delete_track endpoint so that it deletes based on file_path and not title. Updated build script to only write new environment variables if they aren't present and now script unsets these variables after build completetion. Removed unnecessary indents in search_queries file. Updated bad_words check to properly remove titles from the list. Updated ezstream-reread.sh to properly restart station after exhausting all tracks. Updated grab_now_playing.py to print out "No more tracks in playlist" instead of error. Updated update_filename.py to properly format filenames. Fixed a bug in the download script that would incorrectly save the file_path with spaces instead of udnerscores.

+++ scripts/ezstream-reread.sh
@@ -15,12 +15,17 @@ python /stringwave/scripts/remove_whitespaces.py new
 
 find /stringwave/radio/"$station" -name "*.opus" > /stringwave/radio/"$station"/.playlist
 
-ezpid="$(cat /stringwave/.pid-$station)"
-
-# checks if station is run and starts it if it wasn't running, otherwise triggers a reread
-if ps h --pid $ezpid
+if [ -f /stringwave/.pid-$station ]
 then
-    kill -1 $ezpid
+	ezpid="$(cat /stringwave/.pid-$station)"
+	# checks if station is run and starts it if it wasn't running, otherwise triggers a reread
+	if ps h --pid $ezpid
+	then
+	    kill -1 $ezpid
+	else
+	    /stringwave/scripts/ezstream.sh "$station"
+	fi
+
 else
-    /stringwave/scripts/ezstream.sh "$station"
+	/stringwave/scripts/ezstream.sh "$station"
 fi
